+++ skrub/__init__.py
@@ -32,6 +32,7 @@ from ._data_ops import (
 from ._datetime_encoder import DatetimeEncoder
 from ._deduplicate import deduplicate
 from ._drop_uninformative import DropUninformative
+from ._duration_encoder import DurationEncoder
 from ._fuzzy_join import fuzzy_join
 from ._gap_encoder import GapEncoder
 from ._interpolation_joiner import InterpolationJoiner
@@ -67,6 +68,7 @@ __all__ = [
     "TableReport",
     "tabular_pipeline",
     "DatetimeEncoder",
+    "DurationEncoder",
     "ToDatetime",
     "Joiner",
     "fuzzy_join",